Clean up snowflake stage
If an old file with the same name (always the same here) is present on the stage, a new file will not be uploaded.
No message is given in this situation, and it looks like you're uploading a new file but still getting the same data error.
This is an optimization of data, so an actual documented feature, but causing bugs more often because of the default settings.